Cleanup and a Gimenez interpolation bugfix.
